Move some general login things to separate file
